fix: use local envp
