+++ src/exe/exe.h
@@ -41,7 +41,6 @@ int				clean_args(t_cmd_info *cmd);
 int				open_pipe(t_app *app, t_cmd_info *cmd);
 //exe_utils.c
 int				is_valid(t_cmd_info *cmd);
-void			reroute_io(int input, int output);
 void			*free_paths(t_cmd_info *cmd, char **paths,
 					int len, char *cmd_name);
 int				get_paths_len(char **paths);
