+++ src/init/init.c
@@ -11,30 +11,6 @@
 /* ************************************************************************** */
 
 #include "init.h"
-#include "../malloc_list/malloc_list.h"
-#include "../utils/utils.h"
-
-char	**copy_envp(char **envp)
-{
-	int		size;
-	int		i;
-	char	**copy;
-
-	size = 0;
-	while (envp[size])
-		size++;
-	copy = (char **)malloc(sizeof(char *) * (size + 1));
-	if (!copy)
-		return (NULL);
-	i = 0;
-	while (envp[i])
-	{
-		copy[i] = ft_strdup(envp[i]);
-		i++;
-	}
-	copy[i] = NULL;
-	return (copy);
-}
 
 t_app	*init_shell(char *envp[])
 {
