+++ src/init/init.h
@@ -16,6 +16,7 @@
 # include "../includes/types.h"
 # include "../error/error.h"
 # include "../malloc_list/malloc_list.h"
+# include "../utils/utils.h"
 
 t_app			*init_shell(char *envp[]);
 t_parser_info	init_parser_info(int infile, int outfile, char *line);
